+++ Gruntfile.js
@@ -22,7 +22,6 @@ module.exports = function(grunt) {
         sass: 'grunt-sass',
         connect: 'grunt-contrib-connect',
         assemble: 'grunt/assemble/',
-        webpack: 'grunt/webpack/',
         mochaTest: 'grunt-mocha-test',
         open: 'grunt-open'
       }
@@ -41,12 +40,13 @@ module.exports = function(grunt) {
   grunt.registerTask('staging-deploy', [
     'gitinfo',
     'config:staging',
-    'clean:preBuild',
+    'jshint:clientDev',
     'jshint:server',
+    'clean:preBuild',
     'assemble',
+    'handlebars',
     'modernizr',
     'concat',
-    'webpack',
     'sass:prod',
     'autoprefixer',
     'copy',
@@ -58,12 +58,13 @@ module.exports = function(grunt) {
   grunt.registerTask('smartling-staging-deploy', [
     'gitinfo',
     'config:smartlingStaging',
-    'clean:preBuild',
+    'jshint:clientDev',
     'jshint:server',
+    'clean:preBuild',
     'assemble:smartling-staging-deploy',
+    'handlebars',
     'modernizr',
     'concat',
-    'webpack',
     'sass:prod',
     'autoprefixer',
     'copy',
@@ -74,13 +75,14 @@ module.exports = function(grunt) {
 
   var serverTasks = [
     'config:dev',
-    'jshint:test',
+    'jshint:clientDev',
     'jshint:server',
+    'jshint:test',
     'clean:preBuild',
     'assemble',
+    'handlebars',
     'modernizr',
     'concat',
-    'webpack',
     'sass:dev',
     'replace',
     'autoprefixer',
@@ -109,12 +111,13 @@ module.exports = function(grunt) {
 
   grunt.registerTask('build', [
     'config:production',
-    'clean:preBuild',
+    'jshint:clientProd',
     'jshint:server',
+    'clean:preBuild',
     'assemble',
+    'handlebars',
     'modernizr',
     'concat',
-    'webpack',
     'sass:prod',
     'autoprefixer',
     'copy',
@@ -148,13 +151,14 @@ module.exports = function(grunt) {
 
   grunt.registerTask('test', [
     'config:dev',
-    'jshint:test',
+    'jshint:clientProd',
     'jshint:server',
+    'jshint:test',
     'clean:preBuild',
     'assemble',
+    'handlebars',
     'modernizr',
     'concat',
-    'webpack',
     'sass:dev',
     'replace',
     'autoprefixer',
