+++ grunt/concat.js
@@ -42,6 +42,24 @@ var omBundlePaths = [
   'om/global.js'
 ];
 
+var bundlePaths = [
+  'utils/translation.js',
+  'utils/oform_globals.js',
+  'utils/check_complex_password.js',
+  'utils/equal_height_grid.js',
+  'utils/get_url_parameter.js',
+  'utils/uri.js',
+  'utils/trim-mixpanel-cookie.js',
+  'utils/guid_sprintf.js',
+  'utils/form_helper_factory.js',
+  'utils/form_helpers/*.js',
+  'utils/trim_url.js',
+  'global.js',
+  'components/*.js',
+  'services/*.js',
+  '!services/user_state.js'
+];
+
 module.exports = function(grunt, options){
   var lastTarget;
   var processBundleName = function (src, filepath) {
@@ -55,6 +73,19 @@ module.exports = function(grunt, options){
   };
 
   return {
+    namespacePages: {
+      options: {
+        banner: '<%= grunt.config.get("concat_banner") %>',
+        footer: '<%= grunt.config.get("concat_footer") %>',
+        process: function(src, filepath) {
+          return 'var targetName = "' + grunt.task.current.target + '" + "--" +  "' + secondLastSlash(filepath) + '";\n\n' + src;
+        }
+      },
+      src: ['pages/*.js', 'layouts/*.js'],
+      expand: true,
+      cwd: '<%= config.guts %>/assets/js/',
+      dest: '<%= config.dist %>/assets/js/'
+    },
     namespaceOMPages: {
       options: {
         banner: '<%= grunt.config.get("concat_banner") %>',
@@ -98,6 +129,30 @@ module.exports = function(grunt, options){
       flatten: true,
       isFile: true,
       dest: '<%= config.dist %>/assets/js/libraries/jquery-modernizr-om.min.js'
+    },
+    globalBundle: {
+      options: {
+        banner: '<%= grunt.config.get("concat_banner") %>',
+        footer: '<%= grunt.config.get("concat_footer") %>',
+        process: processBundleName
+      },
+      files: {
+        '<%= config.temp %>/assets/js/global.js': makeBundlePaths(bundlePaths)
+      }
+    },
+    concatBundle: {
+      files: {
+        '<%= config.dist %>/assets/js/bundle.js': [
+          '<%= config.bowerDir %>/history.js/scripts/bundled-uncompressed/html4+html5/jquery.history.js',
+          '<%= config.guts %>/assets/js/libraries/handlebars-v1.3.0.js',
+          '<%= config.bowerDir %>/momentjs/moment.js',
+          '<%= config.temp %>/assets/js/handlebarsTemplates.js',
+          '<%= config.bowerDir %>/oform/dist/oForm.min.js',
+          '<%= config.temp %>/assets/js/global.js',
+          '<%= config.guts %>/assets/js/components/oForm-globals.js',
+          '<%= config.bowerDir %>/fitvids/jquery.fitvids.js'
+        ]
+      }
     }
   };
 };
