+++ grunt/jshint.js
@@ -1,10 +1,7 @@
 module.exports = {
   options: {
     trailing: true,
-    forin: false,
-    strict: false,
     curly: true,
-    esnext: true,
     eqeqeq: true,
     indent: 4,
     latedef: true,
@@ -43,6 +40,7 @@ module.exports = {
       unused: true,
       globals: {
         jQuery: false,
+        moment: false,
         $: false,
         Oform: false,
         w: false,
@@ -65,6 +63,7 @@ module.exports = {
       globals: {
         jQuery: false,
         console: false,
+        moment: false,
         _gaq: false,
         $: false,
         Oform: false,
