--- grunt/webpack/loaders/inject-filename.js
@@ -1,14 +1,0 @@
-var loaderUtils = require('loader-utils');
-
-module.exports = function(source, map) {
-  var options = loaderUtils.parseQuery(this.query);
-
-  if (this.cacheable) {
-    this.cacheable();
-  }
-
-  var wrapper = options.banner + options.inject + source + '\n\n' + options.footer;
-  console.log(wrapper);
-
-  this.callback(null, wrapper, map);
-};
