+++ package.json
@@ -23,9 +23,6 @@
   "dependencies": {
     "assemble": "0.6.0-beta.5",
     "assemble-push": "^0.1.1",
-    "babel-core": "^5.0.12",
-    "babel-loader": "^5.0.0",
-    "babel-runtime": "^5.0.12",
     "body-parser": "1.8.3",
     "bower": "1.3.8",
     "chalk": "1.0.0",
@@ -39,6 +36,7 @@
     "grunt-contrib-clean": "0.5.0",
     "grunt-contrib-concat": "0.4.0",
     "grunt-contrib-copy": "0.5.0",
+    "grunt-contrib-handlebars": "0.8.0",
     "grunt-contrib-imagemin": "0.7.1",
     "grunt-contrib-jshint": "0.10.0",
     "grunt-contrib-uglify": "0.5.0",
@@ -50,28 +48,18 @@
     "grunt-text-replace": "0.3.12",
     "grunt-userevvd": "0.1.4",
     "gulp-extname": "0.2.0",
-    "gulp-util": "^3.0.4",
-    "handlebars-loader": "1.0.0",
-    "imports-loader": "0.6.3",
-    "jshint-loader": "0.8.1",
-    "l10n-tools": "git+ssh://git@github.com/optimizely/l10n-tools.git#430ec6c5fa463fdc286b8d26bd6a61e804e1b687",
-    "l10n-tr": "git+ssh://git@github.com/optimizely/l10n-tr.git#4c4175a2c18c4340c4786edb884228c72ccdd14c",
-    "l10n-tr-plugin": "git+ssh://git@github.com/optimizely/l10n-tr-plugin.git#63e9dfbd0bb0ecf90bb65ad3392ab346abe7b729",
+    "handlebars": "3.0.0",
+    "l10n-tools": "git+ssh://git@github.com/optimizely/l10n-tools.git",
     "layout-stack": "^0.1.1",
     "load-grunt-config": "0.16.0",
-    "loader-utils": "^0.2.7",
     "lodash": "^3.2.0",
     "mkdirp": "0.5.0",
-    "moment": "^2.10.2",
     "optimizely-marked": "0.0.4",
     "plasma": "^0.6.1",
     "q": "1.1.2",
-    "script-loader": "0.6.1",
     "through2": "0.6.3",
     "time-grunt": "0.3.2",
-    "uglify-js": "2.4.15",
-    "webpack": "^1.8.4",
-    "webpack-dev-server": "^1.8.0"
+    "uglify-js": "2.4.15"
   },
   "devDependencies": {
     "chai": "2.2.0",
