+++ website-guts/assets/js/pages/webinar.js
@@ -1,6 +1,3 @@
-var moment = require('moment');
-var webinarEventDisplay = require('webinarEventDisplay');
-var webinarEventDetail = require('webinarEventDetail');
 var eventDisplayHTML, templateContext, dateArray, i;
 
 window.optly = window.optly || {};
@@ -79,7 +76,7 @@ for(i = 0; i < dateArray.length; i++){
 
 $(function(){
 
-  eventDisplayHTML = webinarEventDisplay(templateContext);
+  eventDisplayHTML = window.optly.mrkt.templates.webinarEventDisplay(templateContext);
 
   $('#events').html(eventDisplayHTML);
 
@@ -121,7 +118,7 @@ $(function(){
 
     $('.webinar-detail-info').each(function(){
 
-      $(this).html( webinarEventDetail(templateContext.thursdays[index]) );
+      $(this).html( window.optly.mrkt.templates.webinarEventDetail(templateContext.thursdays[index]) );
 
     });
